requirements.txt
python-dotenv~=0.14.0
markovify~=0.8.3
